pew/requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0

